Extract prompt loading logic to prompt-service.js
- Create `src/modules/prompt-service.js` to handle API interaction and caching with stale-while-revalidate strategy.
- Refactor `src/modules/prompt-list.js` to consume `loadPrompts` from the new service.
- Implement background updates via callback in `prompt-list.js`.
- Remove `refreshList` and local `getPromptFolder` from `prompt-list.js`.
- Ensure `createElement` is imported in `prompt-list.js`.
